+++ src/driver/verify.cpp
@@ -150,7 +150,7 @@ static std::vector<argument> run_target(program p,
 bool verify_program(const std::string& name,
                     const program& p,
                     const target& t,
-                    compile_options options,
+                    const compile_options& options,
                     const verify_options& vo,
                     const parameter_map& inputs,
                     verify::tolerance tols)
@@ -181,7 +181,7 @@ bool verify_program(const std::string& name,
 
 void verify_instructions(const program& prog,
                          const target& t,
-                         compile_options options,
+                         const compile_options& options,
                          const verify_options& vo,
                          verify::tolerance tols)
 {
@@ -227,7 +227,7 @@ void verify_instructions(const program& prog,
 static bool verify_reduced(program p,
                            int n,
                            const target& t,
-                           compile_options options,
+                           const compile_options& options,
                            const verify_options& vo,
                            const parameter_map& inputs,
                            verify::tolerance tols)
@@ -251,7 +251,7 @@ static bool verify_reduced(program p,
 
 void verify_reduced_program(const program& p,
                             const target& t,
-                            compile_options options,
+                            const compile_options& options,
                             const verify_options& vo,
                             const parameter_map& inputs,
                             verify::tolerance tols)
@@ -324,7 +324,7 @@ static std::vector<std::size_t> find_trim_instructions(const module& m)
 
 void verify_bisected_program(const program& p,
                              const target& t,
-                             compile_options options,
+                             const compile_options& options,
                              const verify_options& vo,
                              const parameter_map& inputs,
                              verify::tolerance tols)